+++ src/plugins/views/editors/Flow/Components/Links/BaseLink.js
@@ -1,4 +1,5 @@
 import * as d3 from "d3";
+import { LINK_DEPENDENCY } from "../../../../../../utils/Constants";
 import { defaultFunction } from "../../../../../../utils/Utils";
 import { isLinkeable } from "../Nodes/BaseNode/PortValidator";
 import { generatePathPoints } from "./generatePathPoints";
@@ -85,11 +86,19 @@ class BaseLinkStruct {
 }
 
 export default class BaseLink extends BaseLinkStruct {
-  constructor(canvas, src, trg, data, onLinkErrorMouseOver) {
+  constructor(
+    canvas,
+    src,
+    trg,
+    data,
+    flowDebugging = false,
+    onLinkErrorMouseOver = () => defaultFunction("onLinkErrorMouseOver")
+  ) {
     super(canvas, src, trg, data);
     this.object = null;
-    this.onLinkErrorMouseOver =
-      onLinkErrorMouseOver || (() => defaultFunction("onLinkErrorMouseOver"));
+    this.flowDebugging = flowDebugging;
+    this.temporaryDependency = data.Dependency;
+    this.onLinkErrorMouseOver = onLinkErrorMouseOver;
 
     this.initialize();
   }
@@ -134,7 +143,7 @@ export default class BaseLink extends BaseLinkStruct {
       .attr("d", () => {
         return this.calculateLine(pathPoints);
       })
-      .attr("stroke", this.style.stroke.default)
+      .attr("stroke", this.getStrokeColor())
       .attr("stroke-width", this.style.stroke.width)
       .attr("pointer-events", "visibleStroke")
       .attr("fill-opacity", "0");
@@ -161,7 +170,7 @@ export default class BaseLink extends BaseLinkStruct {
    */
   refreshPath = () => {
     const { pathPoints } = this;
-    this.object.select("path").attr("d", () => {
+    this.object.select("path.nodeLink").attr("d", () => {
       return this.calculateLine(pathPoints);
     });
     return this;
@@ -200,8 +209,8 @@ export default class BaseLink extends BaseLinkStruct {
    * @private
    */
   styleMouseOver = () => {
-    this.path
-      .attr("stroke", this.style.stroke.over)
+    this.changeStrokeColor()
+      .path.attr("stroke-width", this.style.stroke.overWidth)
       .attr("marker-mid", `url(#${this.canvas.containerId}-markerselected)`);
   };
 
@@ -209,11 +218,8 @@ export default class BaseLink extends BaseLinkStruct {
    * @private
    */
   styleMouseOut = () => {
-    this.path
-      .attr(
-        "stroke",
-        this.error ? this.style.stroke.warning : this.style.stroke.default
-      )
+    this.changeStrokeColor()
+      .path.attr("stroke-width", this.style.stroke.width)
       .attr("marker-mid", null);
   };
 
@@ -318,6 +324,23 @@ export default class BaseLink extends BaseLinkStruct {
     this.object.attr("visibility", this.visible ? "visible" : "hidden");
   }
 
+  setTemporaryDependency(dependecy) {
+    this.temporaryDependency = dependecy;
+    return this;
+  }
+
+  getStrokeColor() {
+    const type = `dependency_${
+      this.temporaryDependency ?? this.data.Dependency
+    }`;
+    const strokeColor = this.style.stroke.default;
+
+    if (this.flowDebugging && baseLinkStyles[type])
+      return baseLinkStyles[type].color;
+
+    return strokeColor;
+  }
+
   onSelected = selected => {
     this._isSelected = selected;
     if (selected) this.styleMouseOver();
@@ -340,6 +363,14 @@ export default class BaseLink extends BaseLinkStruct {
   updateError = error => {
     this.error = error;
     this.styleMouseOut();
+  };
+
+  changeStrokeColor = () => {
+    this.path.attr(
+      "stroke",
+      this.error ? this.style.stroke.warning : this.getStrokeColor()
+    );
+    return this;
   };
 
   /**
@@ -424,7 +455,12 @@ export default class BaseLink extends BaseLinkStruct {
     });
   }
 
-  static parseLink({ id, From, To, Dependency }) {
+  static parseLink({
+    id,
+    From,
+    To,
+    Dependency = LINK_DEPENDENCY.ALL_DEPENDENCIES.VALUE
+  }) {
     const [sourceNode, sourcePort, sourceFullPath] = BaseLink.getNodePort(From);
     const [targetNode, targetPort, targetFullPath] = BaseLink.getNodePort(To);