+++ src/utils/Workspace.js
@@ -1,4 +1,4 @@
-import { Authentication } from "@mov-ai/mov-fe-lib-core";
+import { User } from "@mov-ai/mov-fe-lib-core";
 import { DOCK_POSITIONS, DEFAULT_LAYOUT, DEFAULT_TABS } from "./Constants";
 import LocalStorage from "./LocalStorage";
 
@@ -6,9 +6,10 @@ class Workspace {
   constructor() {
     if (instance) return instance;
     instance = this;
+    this.user = new User();
 
     const APP_NAME = "movai-ide-ce";
-    const USER_NAME = Authentication.getTokenData().message.name ?? "";
+    const USER_NAME = this.user.getUsername() ?? "";
 
     this.storage = new LocalStorage();
     this.TABS_KEY = `movai.${USER_NAME}.${APP_NAME}.tabs`;
@@ -16,6 +17,7 @@ class Workspace {
     this.LAYOUT_KEY = `movai.${USER_NAME}.${APP_NAME}.layout`;
     this.SELECTED_ROBOT_KEY = `movai.${USER_NAME}.${APP_NAME}.selectedRobot`;
     this.RECENT_DOCUMENTS_KEY = `movai.${USER_NAME}.${APP_NAME}.recentDocuments`;
+    this.DEBUGGING_FLOW_KEY = `movai.${USER_NAME}.${APP_NAME}.flowIsDebugging`;
     this.layoutAndTabs = this.getLayoutAndTabs();
     this.layout = this.layoutAndTabs[0];
     this.tabs = this.layoutAndTabs[1];
@@ -123,6 +125,28 @@ class Workspace {
     }
     // Return formatted tabStack
     return tabStack;
+  }
+
+  //========================================================================================
+  /*                                                                                      *
+   *                               Flow Debug : Used in Flow                              *
+   *                                                                                      */
+  //========================================================================================
+
+  /**
+   * Attempts to get the stored flowDebugging value, false is default
+   * @returns {Boolean} : Is the flow debugging?
+   */
+  getFlowIsDebugging() {
+    return this.storage.get(this.DEBUGGING_FLOW_KEY) ?? false;
+  }
+
+  /**
+   * Attempts to set the stored flowDebugging value, sets false if nothing is passed
+   * @param {Boolean} isFlowDebugging : Is the flow debugging?
+   */
+  setFlowIsDebugging(isFlowDebugging) {
+    this.storage.set(this.DEBUGGING_FLOW_KEY, isFlowDebugging ?? false);
   }
 
   //========================================================================================